+++ config/initializers/bugsnag.rb
@@ -1,3 +1,3 @@
 Bugsnag.configure do |config|
-  config.api_key = "dcc202101f7a3935202fe585eec010c8"
+  config.api_key = Rails.application.secrets.bugsnag_api_key
 end