+++ spec/factories/params_factory.rb
@@ -1,11 +1,11 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :user_params, class: Hash do
     skip_create
 
-    name 'John Snow'
-    email 'john@snow.com'
-    provider 'spotify'
-    uid '456'
+    name { 'John Snow' }
+    email { 'john@snow.com' }
+    provider { 'spotify' }
+    uid { '456' }
 
     initialize_with do
       {